Add desktop websocket support and json wrapper

+++ src/fetch_url/fetch_url.c
@@ -107,7 +107,7 @@ fetch_url_op_t *fetch_url_async(const char *url, int timeout_ms)
     curl_easy_setopt(curl, CURLOPT_URL, url);
     curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, write_callback);
     curl_easy_setopt(curl, CURLOPT_WRITEDATA, &context);
-    curl_easy_setopt(curl, CURLOPT_TIMEOUT_MS, timeout_ms);
+    curl_easy_setopt(curl, CURLOPT_TIMEOUT_MS, (long)timeout_ms);
 
     res = curl_easy_perform(curl);
     if (res == CURLE_OK)
